@@ -5,7 +5,7 @@ authors = ["Martin Thomson <mt@lowentropy.net>",
 	"Dragana Damjanovic <dragana.damjano@gmail.com>",
 	"Andy Grover <agrover@mozilla.com>"]
 edition = "2018"
-rust-version = "1.57.0"
+rust-version = "1.65.0"
 license = "MIT/Apache-2.0"
 
 [dependencies]
