@@ -3,7 +3,7 @@ name = "neqo-common"
 version = "0.6.4"
 authors = ["Bobby Holley <bobbyholley@gmail.com>"]
 edition = "2018"
-rust-version = "1.57.0"
+rust-version = "1.65.0"
 license = "MIT/Apache-2.0"
 build = "build.rs"
 
