@@ -3,7 +3,7 @@ name = "neqo-crypto"
 version = "0.6.4"
 authors = ["Martin Thomson <mt@lowentropy.net>"]
 edition = "2018"
-rust-version = "1.57.0"
+rust-version = "1.65.0"
 build = "build.rs"
 license = "MIT/Apache-2.0"
 
