@@ -3,7 +3,7 @@ name = "neqo-interop"
 version = "0.6.4"
 authors = ["EKR <ekr@rtfm.com>"]
 edition = "2018"
-rust-version = "1.57.0"
+rust-version = "1.65.0"
 license = "MIT/Apache-2.0"
 
 [dependencies]
