@@ -176,7 +176,7 @@ fn dynamic_link_both(extra_libs: &[&str]) {
         &["plds4", "plc4", "nspr4"]
     };
     for lib in nspr_libs.iter().chain(extra_libs) {
-        println!("cargo:rustc-link-lib=dylib={}", lib);
+        println!("cargo:rustc-link-lib=dylib={lib}");
     }
 }
 
@@ -202,7 +202,7 @@ fn static_link() {
         static_libs.push("sqlite");
     }
     for lib in static_libs {
-        println!("cargo:rustc-link-lib=static={}", lib);
+        println!("cargo:rustc-link-lib=static={lib}");
     }
 
     // Dynamic libs that aren't transitively included by NSS libs.
@@ -232,7 +232,7 @@ fn build_bindings(base: &str, bindings: &Bindings, flags: &[String], gecko: bool
     let header = header_path.to_str().unwrap();
     let out = PathBuf::from(env::var("OUT_DIR").unwrap()).join(String::from(base) + ".rs");
 
-    println!("cargo:rerun-if-changed={}", header);
+    println!("cargo:rerun-if-changed={header}");
 
     let mut builder = Builder::default().header(header);
     builder = builder.generate_comments(false);
