@@ -443,10 +443,7 @@ impl Http3Client {
             return Err(Error::InvalidState);
         }
         let mut dec = Decoder::from(token.as_ref());
-        let settings_slice = match dec.decode_vvec() {
-            Some(v) => v,
-            None => return Err(Error::InvalidResumptionToken),
-        };
+        let Some(settings_slice) = dec.decode_vvec() else { return Err(Error::InvalidResumptionToken) };
         qtrace!([self], "  settings {}", hex_with_len(settings_slice));
         let mut dec_settings = Decoder::from(settings_slice);
         let mut settings = HSettings::default();
@@ -870,7 +867,7 @@ impl Http3Client {
             }
             Err(e) => {
                 qinfo!([self], "Connection error: {}.", e);
-                self.close(now, e.code(), &format!("{}", e));
+                self.close(now, e.code(), &format!("{e}"));
                 true
             }
             _ => false,
