@@ -179,7 +179,7 @@ mod tests {
             for tail in 1..db.len() {
                 let split = db.len() - tail;
                 let mut dv = Decoder::from(&db.as_ref()[0..split]);
-                eprintln!("  split at {}: {:?}", split, dv);
+                eprintln!("  split at {split}: {dv:?}");
 
                 // Clone the basic decoder for each iteration of the loop.
                 let mut dec = decoder.clone();
@@ -193,7 +193,7 @@ mod tests {
                     assert_eq!(res, None);
                     assert!(dec.min_remaining() > 0);
                     let mut dv = Decoder::from(&db.as_ref()[split..]);
-                    eprintln!("  split remainder {}: {:?}", split, dv);
+                    eprintln!("  split remainder {split}: {dv:?}");
                     res = dec.consume(&mut dv);
                     assert_eq!(dv.remaining(), 1);
                 }
@@ -245,7 +245,7 @@ mod tests {
         for tail in 1..db.len() {
             let split = db.len() - tail;
             let mut dv = Decoder::from(&db.as_ref()[0..split]);
-            eprintln!("  split at {}: {:?}", split, dv);
+            eprintln!("  split at {split}: {dv:?}");
 
             // Clone the basic decoder for each iteration of the loop.
             let mut dec = decoder.clone();
@@ -257,7 +257,7 @@ mod tests {
                 assert!(!res);
                 assert!(dec.min_remaining() > 0);
                 let mut dv = Decoder::from(&db.as_ref()[split..]);
-                eprintln!("  split remainder {}: {:?}", split, dv);
+                eprintln!("  split remainder {split}: {dv:?}");
                 res = dec.consume(&mut dv);
                 assert_eq!(dv.remaining(), 1);
             }
