@@ -42,10 +42,8 @@ fn stapled_ocsp_responses(fd: *mut PRFileDesc) -> Option<Vec<Vec<u8>>> {
     match NonNull::new(ocsp_nss as *mut SECItemArray) {
         Some(ocsp_ptr) => {
             let mut ocsp_helper: Vec<Vec<u8>> = Vec::new();
-            let len = if let Ok(l) = isize::try_from(unsafe { ocsp_ptr.as_ref().len }) {
-                l
-            } else {
-                qerror!([format!("{:p}", fd)], "Received illegal OSCP length");
+            let Ok(len) = isize::try_from(unsafe { ocsp_ptr.as_ref().len }) else {
+                qerror!([format!("{fd:p}")], "Received illegal OSCP length");
                 return None;
             };
             for idx in 0..len {
