@@ -73,7 +73,7 @@ impl std::error::Error for Error {
 
 impl std::fmt::Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "Error: {:?}", self)
+        write!(f, "Error: {self:?}")
     }
 }
 
