@@ -63,10 +63,7 @@ impl ControlStreamLocal {
     ) -> Res<()> {
         // send all necessary priority updates
         while let Some(update_id) = self.outstanding_priority_update.pop_front() {
-            let update_stream = match recv_conn.get_mut(&update_id) {
-                Some(update_stream) => update_stream,
-                None => continue,
-            };
+            let Some(update_stream) = recv_conn.get_mut(&update_id) else { continue };
 
             // can assert and unwrap here, because priority updates can only be added to
             // HttpStreams in [Http3Connection::queue_update_priority}
