@@ -42,10 +42,7 @@ impl FrameDecoder<WebTransportFrame> for WebTransportFrame {
                 }
                 let error =
                     u32::try_from(dec.decode_uint(4).ok_or(Error::HttpMessageError)?).unwrap();
-                let message = match String::from_utf8(dec.decode_remainder().to_vec()) {
-                    Ok(s) => s,
-                    Err(_) => return Err(Error::HttpMessageError),
-                };
+                let Ok(message) = String::from_utf8(dec.decode_remainder().to_vec()) else { return Err(Error::HttpMessageError) };
                 Ok(Some(WebTransportFrame::CloseSession { error, message }))
             } else {
                 Ok(None)
