@@ -408,7 +408,7 @@ impl ::std::error::Error for Error {
 
 impl ::std::fmt::Display for Error {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
-        write!(f, "HTTP/3 error: {:?}", self)
+        write!(f, "HTTP/3 error: {self:?}")
     }
 }
 
