@@ -101,7 +101,7 @@ impl ::std::error::Error for Error {
 
 impl ::std::fmt::Display for Error {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
-        write!(f, "QPACK error: {:?}", self)
+        write!(f, "QPACK error: {self:?}")
     }
 }
 
